SentimentAnalysis/requirements.txt
pandas==1.5.3
numpy==1.25.0
scikit-learn==1.2.2
scipy==1.10.1
matplotlib==3.8.3
joblib==1.3.2
gensim==4.3.2
nltk==3.8.1
Flask==3.0.3
contractions==0.1.73
pytest==7.4.0
pytest-cov==4.1.0
PyYAML==6.0
requests==2.31.0
